refactor: move GlobalStyles to styles
