+++ src/styles/index.ts
@@ -1,1 +1,2 @@
 export { colors } from './colors';
+export { GlobalStyle } from './globalStyle';
